scripts/websearch-custom-endpoints/requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
# Note: Install OpenAI Agents SDK separately:
# pip install git+https://github.com/openai/openai-agents-sdk